# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vendor Control Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List all vendors
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a new vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{vendorId}:
    parameters:
      - name: vendorId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get a single vendor with stats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorDetail"
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor info (plan, phone, name, bank)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVendorBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete a vendor
      responses:
        "204":
          description: Deleted
...
  /vendors/{vendorId}/promotions:
    parameters:
      - name: vendorId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: listVendorPromotions
      tags: [promotions]
      summary: List promotions for a vendor
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Promotion"
    post:
      operationId: createVendorPromotion
      tags: [promotions]
      summary: Create a promotion (Pro feature)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromotionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
  /vendors/{vendorId}/promotions/{promotionId}:
    parameters:
      - name: vendorId
        in: path
        required: true
        schema: { type: string }
      - name: promotionId
        in: path
        required: true
        schema: { type: string }
    patch:
      operationId: updateVendorPromotion
      tags: [promotions]
      summary: Update a promotion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePromotionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
    delete:
      operationId: deleteVendorPromotion
      tags: [promotions]
      summary: Delete a promotion
      responses:
        "204":
          description: Deleted
  /vendors/{vendorId}/broadcasts:
    parameters:
      - name: vendorId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: listVendorBroadcasts
      tags: [broadcasts]
      summary: List past broadcasts
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Broadcast"
    post:
      operationId: sendVendorBroadcast
      tags: [broadcasts]
      summary: Send a broadcast to recent customers (Pro feature)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBroadcastBody"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Broadcast"
  /vendors/{vendorId}/follow-ups/run:
    parameters:
      - name: vendorId
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: runVendorFollowUps
      tags: [broadcasts]
      summary: Send follow-up reminders to customers with stalled orders (Pro feature)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUpResult"
...
components:
  schemas:
...
    Plan:
      type: string
      enum: [starter, pro]
    Vendor:
      type: object
      required: [id, name, phoneNumber, plan, botEnabled, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        phoneNumber: { type: string }
        adminNumber: { type: string, nullable: true }
        phoneNumberId: { type: string, nullable: true }
        botNumber: { type: string, nullable: true }
        plan: { $ref: "#/components/schemas/Plan" }
        botEnabled: { type: boolean }
        bankName: { type: string, nullable: true }
        bankAccountNumber: { type: string, nullable: true }
        bankAccountHolder: { type: string, nullable: true }
        currency: { type: string }
        welcomeMessage: { type: string, nullable: true }
        followUpsEnabled: { type: boolean }
        createdAt: { type: string, format: date-time }
    VendorDetail:
      allOf:
        - $ref: "#/components/schemas/Vendor"
        - type: object
          required: [stats]
          properties:
            stats:
              $ref: "#/components/schemas/VendorStats"
    VendorStats:
      type: object
      required:
        - totalOrders
        - pendingOrders
        - confirmedOrders
        - paidOrders
        - revenue
        - menuItems
        - customers
        - openConversations
      properties:
        totalOrders: { type: integer }
        pendingOrders: { type: integer }
        confirmedOrders: { type: integer }
        paidOrders: { type: integer }
        revenue: { type: number }
        menuItems: { type: integer }
        customers: { type: integer }
        openConversations: { type: integer }
    CreateVendorBody:
      type: object
      required: [name, phoneNumber, plan]
      properties:
        name: { type: string, minLength: 1 }
        phoneNumber: { type: string, minLength: 4 }
        adminNumber: { type: string }
        phoneNumberId: { type: string }
        botNumber: { type: string }
        plan: { $ref: "#/components/schemas/Plan" }
        bankName: { type: string }
        bankAccountNumber: { type: string }
        bankAccountHolder: { type: string }
        currency: { type: string }
        welcomeMessage: { type: string }
    UpdateVendorBody:
      type: object
      properties:
        name: { type: string }
        phoneNumber: { type: string }
        adminNumber: { type: string }
        phoneNumberId: { type: string }
        botNumber: { type: string }
        plan: { $ref: "#/components/schemas/Plan" }
        botEnabled: { type: boolean }
        followUpsEnabled: { type: boolean }
        bankName: { type: string }
        bankAccountNumber: { type: string }
        bankAccountHolder: { type: string }
        currency: { type: string }
        welcomeMessage: { type: string }
...
    Promotion:
      type: object
      required: [id, vendorId, title, active, createdAt]
      properties:
        id: { type: string }
        vendorId: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
    CreatePromotionBody:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        active: { type: boolean }
    UpdatePromotionBody:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        active: { type: boolean }
    Broadcast:
      type: object
      required: [id, vendorId, message, recipientCount, sentAt]
      properties:
        id: { type: string }
        vendorId: { type: string }
        message: { type: string }
        recipientCount: { type: integer }
        sentAt: { type: string, format: date-time }
    CreateBroadcastBody:
      type: object
      required: [message]
      properties:
        message: { type: string, minLength: 1 }
        sinceDays:
          type: integer
          minimum: 1
          description: "Only customers seen in the last N days. Default 30."
    FollowUpResult:
      type: object
      required: [reminded, broadcastId]
      properties:
        reminded: { type: integer }
        broadcastId: { type: string, nullable: true }