# lib/api-zod/src/generated/types/createBroadcastBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vendor Control Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBroadcastBody {
  /** @minLength 1 */
  message: string;
  /**
   * Only customers seen in the last N days. Default 30.
   * @minimum 1
   */
  sinceDays?: number;
}