# lib/api-zod/src/generated/types/followUpResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vendor Control Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface FollowUpResult {
  reminded: number;
  broadcastId: string | null;
}