# lib/api-zod/src/generated/types/promotion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vendor Control Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface Promotion {
  id: string;
  vendorId: string;
  title: string;
  description?: string | null;
  active: boolean;
  createdAt: Date;
}