# lib/api-zod/src/generated/types/updateVendorBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vendor Control Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { Plan } from "./plan";

export interface UpdateVendorBody {
  name?: string;
  phoneNumber?: string;
  adminNumber?: string;
  phoneNumberId?: string;
  botNumber?: string;
  plan?: Plan;
  botEnabled?: boolean;
  followUpsEnabled?: boolean;
  bankName?: string;
  bankAccountNumber?: string;
  bankAccountHolder?: string;
  currency?: string;
  welcomeMessage?: string;
}