# lib/api-zod/src/generated/types/vendor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vendor Control Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { Plan } from "./plan";

export interface Vendor {
  id: string;
  name: string;
  phoneNumber: string;
  adminNumber?: string | null;
  phoneNumberId?: string | null;
  botNumber?: string | null;
  plan: Plan;
  botEnabled: boolean;
  bankName?: string | null;
  bankAccountNumber?: string | null;
  bankAccountHolder?: string | null;
  currency?: string;
  welcomeMessage?: string | null;
  followUpsEnabled?: boolean;
  createdAt: Date;
}